+++ include/api.h
@@ -55,26 +55,27 @@
 #include "pros/misc.h"
 #include "pros/motors.h"
 #include "pros/optical.h"
-#include "pros/rtos.h"
 #include "pros/rotation.h"
+#include "pros/rtos.h"
 #include "pros/screen.h"
 #include "pros/vision.h"
 
 #ifdef __cplusplus
 #include "pros/adi.hpp"
 #include "pros/colors.hpp"
+#include "pros/device.hpp"
 #include "pros/distance.hpp"
 #include "pros/gps.hpp"
 #include "pros/imu.hpp"
+#include "pros/link.hpp"
 #include "pros/misc.hpp"
+#include "pros/motor_group.hpp"
 #include "pros/motors.hpp"
 #include "pros/optical.hpp"
 #include "pros/rotation.hpp"
 #include "pros/rtos.hpp"
 #include "pros/screen.hpp"
 #include "pros/vision.hpp"
-#include "pros/link.hpp"
-#include "pros/device.hpp"
 #endif
 
 #endif  // _PROS_API_H_
