+++ src/devices/vdml_motors.c
@@ -15,15 +15,14 @@
 #include <stdio.h>
 
 #include "kapi.h"
+#include "pros/colors.h"
 #include "pros/motors.h"
 #include "v5_api.h"
 #include "vdml/registry.h"
 #include "vdml/vdml.h"
-#include "pros/colors.h"
 
 #define MOTOR_MOVE_RANGE 127
 #define MOTOR_VOLTAGE_RANGE 12000
-
 
 // Movement functions
 
@@ -34,10 +33,10 @@ int32_t motor_move(int8_t port, int32_t voltage) {
 		voltage = -127;
 	}
 
-	if(port < 0) {
+	if (port < 0) {
 		voltage = -voltage;
 	}
-	
+
 	port = abs(port);
 	// Remap the input voltage range to the motor voltage
 	// scale to [-127, 127] -> [-12000, 12000]
@@ -53,7 +52,7 @@ int32_t motor_brake(int8_t port) {
 int32_t motor_move_absolute(int8_t port, const double position, const int32_t velocity) {
 	uint8_t abs_port = abs(port);
 	claim_port_i(abs_port - 1, E_DEVICE_MOTOR);
-	if(port < 0) velocity = -velocity;
+	// if(port < 0) velocity = -velocity;
 	vexDeviceMotorAbsoluteTargetSet(device->device_info, position, velocity);
 	return_port(abs_port - 1, PROS_SUCCESS);
 }
@@ -61,7 +60,7 @@ int32_t motor_move_absolute(int8_t port, const double position, const int32_t ve
 int32_t motor_move_relative(int8_t port, const double position, const int32_t velocity) {
 	uint8_t abs_port = abs(port);
 	claim_port_i(abs_port - 1, E_DEVICE_MOTOR);
-	if(port < 0) velocity = -velocity;
+	// if(port < 0) velocity = -velocity;
 	vexDeviceMotorRelativeTargetSet(device->device_info, position, velocity);
 	return_port(abs_port - 1, PROS_SUCCESS);
 }
@@ -69,7 +68,7 @@ int32_t motor_move_relative(int8_t port, const double position, const int32_t ve
 int32_t motor_move_velocity(int8_t port, const int32_t velocity) {
 	uint8_t abs_port = abs(port);
 	claim_port_i(abs_port - 1, E_DEVICE_MOTOR);
-	if(port < 0) velocity = -velocity;
+	// if(port < 0) velocity = -velocity;
 	vexDeviceMotorVelocitySet(device->device_info, velocity);
 	return_port(abs_port - 1, PROS_SUCCESS);
 }
@@ -77,7 +76,7 @@ int32_t motor_move_velocity(int8_t port, const int32_t velocity) {
 int32_t motor_move_voltage(int8_t port, const int32_t voltage) {
 	uint8_t abs_port = abs(port);
 	claim_port_i(abs_port - 1, E_DEVICE_MOTOR);
-	if(port < 0) voltage = -voltage;
+	// if(port < 0) voltage = -voltage;
 	vexDeviceMotorVoltageSet(device->device_info, voltage);
 	return_port(abs_port - 1, PROS_SUCCESS);
 }
@@ -85,7 +84,7 @@ int32_t motor_move_voltage(int8_t port, const int32_t voltage) {
 int32_t motor_modify_profiled_velocity(int8_t port, const int32_t velocity) {
 	uint8_t abs_port = abs(port);
 	claim_port_i(abs_port - 1, E_DEVICE_MOTOR);
-	if(port < 0) velocity = -velocity;
+	// if(port < 0) velocity = -velocity;
 	vexDeviceMotorVelocityUpdate(device->device_info, velocity);
 	return_port(abs_port - 1, PROS_SUCCESS);
 }
@@ -147,8 +146,6 @@ int32_t motor_is_over_temp(int8_t port) {
 	int rtn = vexDeviceMotorOverTempFlagGet(device->device_info);
 	return_port(port - 1, rtn);
 }
-
-
 
 uint32_t motor_get_faults(int8_t port) {
 	port = abs(port);
@@ -248,13 +245,20 @@ int32_t motor_set_gearing(int8_t port, const motor_gearset_e_t gearset) {
 	claim_port_i(port - 1, E_DEVICE_MOTOR);
 	motor_gearset_e_t final_gearset = gearset;
 	// in the case that the motor gearset is actually a color, we need to handle this too:
-	if(gearset > E_MOTOR_GEAR_BLUE) {
-		switch(gearset) {
-			case COLOR_RED: 	final_gearset = E_MOTOR_GEAR_RED; break;
-			case COLOR_GREEN: 	final_gearset = E_MOTOR_GEAR_GREEN; break;
-			case COLOR_BLUE: 	final_gearset = E_MOTOR_GEAR_BLUE; break;
-			default: 		 	errno = EINVAL; 
-								return_port(port - 1, PROS_ERR);
+	if (gearset > E_MOTOR_GEAR_BLUE) {
+		switch (gearset) {
+			case COLOR_RED:
+				final_gearset = E_MOTOR_GEAR_RED;
+				break;
+			case COLOR_GREEN:
+				final_gearset = E_MOTOR_GEAR_GREEN;
+				break;
+			case COLOR_BLUE:
+				final_gearset = E_MOTOR_GEAR_BLUE;
+				break;
+			default:
+				errno = EINVAL;
+				return_port(port - 1, PROS_ERR);
 		}
 	}
